requirements.txt
python
torch
termcolor
wandb
hydra-core
pandas
attrdictionary
matplotlib
ipykernel
scikit-learn
scipy
